Divergence review using surface minimum

+++ premiumMarkets/pm-core/src/main/java/com/finance/pms/events/operations/conditional/HigherLowCondition.java
@@ -54,7 +54,7 @@ public class HigherLowCondition extends HighsAndLowsCondition {
 	@Override
 	public Boolean conditionCheck(Comparable... ops) {
 		return highLowSolver.higherLow(
-				(SortedMap<Integer, Double>) ops[0], (Integer) ops[1], (Integer) ops[2], (SortedMap<Integer, Double>) ops[3], ((Line<Integer, Double>) ops[4]),
+				(SortedMap<Integer, Double>) ops[0], (Integer) ops[1], (Double) ops[2], (SortedMap<Integer, Double>) ops[3], ((Line<Integer, Double>) ops[4]),
 				(Double) ops[5], (Double) ops[6], (Double) ops[7], (Double) ops[8],
 				(Double) ops[9], (Double) ops[10]);
 	}
